+++ js/tone_row.js
@@ -11,18 +11,18 @@ it also gets displayed to the screen
 var _flats_on = false
 
 //remove selected note from alphabet array
-function remove_note(note){
-	var index = _alphabet.indexOf(note)	//address of note
+function remove_note(note, alphabet){
+	var index = alphabet.indexOf(note)	//address of note
 	if (index > -1)
-		_alphabet.splice(index, 1)
+		alphabet.splice(index, 1)
 	//console.log(note + "%c has been removed", "color: #9A0000")
 }
 
 //returns current element id string
 function get_element_id(array_length){
-	str1 = "note_"
-	num1 = 12-array_length
-	return element_id =  str1.concat(num1.toString())
+	str = "note_"
+	num = 12-array_length
+	return element_id =  str.concat(num.toString())
 }
 
 //switches based on user input to which accidental array gets used
@@ -35,46 +35,52 @@ function toggle_accidental(toggle){
 
 //loads array depending on user's accidental bool preference
 function load_array(){
-	console.clear()	//only show the current tone row log
-	
+	//define both types of musical alphabets
     var sharps = ['A', 'A#', 'B', 'C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#']
     var flats = ['A', 'Bb', 'B', 'C', 'Db', 'D', 'Eb', 'E', 'F', 'Gb', 'G', 'Ab']
-	_alphabet = []	//unnecessarily cleans array
-	_tone_row = []	//necessarily cleans array
 	
-	if(_flats_on){
+	//resets arrays so they'll be clean
+	_alphabet = []
+	_tone_row = []
+	
+	//if toggle is set to flat, load flats
+	if(_flats_on)
 		_alphabet = flats.concat()
-		_alphabet_copy = flats.concat()
-	}
-	else{
+	else
 		_alphabet = sharps.concat()
-		_alphabet_copy = sharps.concat()
-	}
 }
 
 //main function for tone_row
 function tone_row(){
-	load_array()	//puts either sharps or flats in array
-
-	//if custom toggle is check (currently breaks program if true)
-	if(custom == false){
-		while(_alphabet.length){	//loop 12 times
-
-			//random int from 0 to length of array
-			var random_address = Math.floor((Math.random()*_alphabet.length))	
-			var note = _alphabet[random_address]	//select note based on address
-			var element_id = get_element_id(_alphabet.length)	//create and store element id
-
-			document.getElementById(element_id).innerHTML = note	//send note to html
-			_tone_row.push(note)	//store note in final array to use later
-
-			remove_note(note)	//remove note from the alphabet array
-		}
-	}
+	console.log("%ctone_row(): \t%cstart", "color: blue; font-weight:bold;", "color: orange")
+	
+	//this loads the correct musical alphabet
+	//depening on the settings, alphabet can be sharps or flats
+	load_array()
+	
+	//makes a local copy of the alphabet to use
+	//after each loop, arrays shrinks by one
+	var shrinking_alphabet = _alphabet.slice()
 	
 	//if user input a custom row
 	if(custom == true){
 		//take user input and store it in the _tone_row
-		console.log("%c::done::", "color: #009A00")
+		console.log("%c::custom row?::", "color: #009A00")
 	}
+	else{
+		for(var i = 12 ; i > 0 ; i--){
+			//random int from 0 to length of array
+			var random_address = Math.floor((Math.random()*i))	
+			var note = shrinking_alphabet[random_address]	//select note based on address
+			var element_id = get_element_id(i)	//create and store element id
+
+			//send note to html
+			document.getElementById(element_id).innerHTML = note	
+			_tone_row.push(note)	//store note in final array to use later
+
+			remove_note(note, shrinking_alphabet)	//remove note from the alphabet array
+		}
+	}
+	
+	console.log("%ctone_row(): \t%cdone", "color: blue; font-weight:bold;", "color: limegreen")
 }